01_basico/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==3.0.7
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
pycodestyle==2.9.1
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
zipp==3.8.1
